+++ tests/test_all.py
@@ -25,6 +25,7 @@ if __name__ == '__main__':
 
         tests.addTests(loader.discover(start_dir=tests_dir, pattern="test_etree.py"))
         tests.addTests(loader.discover(start_dir=tests_dir, pattern="test_helpers.py"))
+        tests.addTests(loader.discover(start_dir=tests_dir, pattern="test_cli.py"))
         tests.addTests(loader.discover(start_dir=tests_dir, pattern="test_converters.py"))
         tests.addTests(loader.discover(start_dir=tests_dir, pattern="test_regex.py"))
         tests.addTests(loader.discover(start_dir=tests_dir, pattern="test_resources.py"))