+++ xmlschema/__init__.py
@@ -9,15 +9,15 @@
 # @author Davide Brunato <brunato@sissa.it>
 #
 from . import limits
-from .exceptions import XMLSchemaException, XMLSchemaRegexError, XMLSchemaURLError, \
-    XMLSchemaNamespaceError
+from .exceptions import XMLSchemaException, XMLSchemaRegexError, \
+    XMLSchemaResourceError, XMLSchemaNamespaceError
 from .etree import etree_tostring
 from .resources import normalize_url, normalize_locations, fetch_resource, \
     fetch_namespaces, fetch_schema_locations, fetch_schema, XMLResource
 from .xpath import ElementPathMixin
 from .converters import (
     ElementData, XMLSchemaConverter, UnorderedConverter, ParkerConverter,
-    BadgerFishConverter, AbderaConverter, JsonMLConverter
+    BadgerFishConverter, AbderaConverter, JsonMLConverter, ColumnarConverter
 )
 from .documents import validate, is_valid, iter_errors, to_dict, to_json, from_json
 
@@ -26,10 +26,11 @@ from .validators import (
     XMLSchemaModelError, XMLSchemaModelDepthError, XMLSchemaValidationError,
     XMLSchemaDecodeError, XMLSchemaEncodeError, XMLSchemaChildrenValidationError,
     XMLSchemaIncludeWarning, XMLSchemaImportWarning, XMLSchemaTypeTableWarning,
-    XsdGlobals, XMLSchemaBase, XMLSchema, XMLSchema10, XMLSchema11
+    XsdGlobals, XMLSchemaBase, XMLSchema, XMLSchema10, XMLSchema11,
+    XsdComponent, XsdType, XsdElement, XsdAttribute
 )
 
-__version__ = '1.1.3'
+__version__ = '1.2.0'
 __author__ = "Davide Brunato"
 __contact__ = "brunato@sissa.it"
 __copyright__ = "Copyright 2016-2020, SISSA"
@@ -38,15 +39,17 @@ __status__ = "Production/Stable"
 
 
 __all__ = [
-    'limits', 'XMLSchemaException', 'XMLSchemaRegexError', 'XMLSchemaURLError',
+    'limits', 'XMLSchemaException', 'XMLSchemaRegexError', 'XMLSchemaResourceError',
     'XMLSchemaNamespaceError', 'etree_tostring', 'normalize_url', 'normalize_locations',
     'fetch_resource', 'fetch_namespaces', 'fetch_schema_locations', 'fetch_schema',
     'XMLResource', 'ElementPathMixin', 'ElementData', 'XMLSchemaConverter',
     'UnorderedConverter', 'ParkerConverter', 'BadgerFishConverter', 'AbderaConverter',
-    'JsonMLConverter', 'validate', 'is_valid', 'iter_errors', 'to_dict', 'to_json',
-    'from_json', 'XMLSchemaValidatorError', 'XMLSchemaParseError', 'XMLSchemaNotBuiltError',
-    'XMLSchemaModelError', 'XMLSchemaModelDepthError', 'XMLSchemaValidationError',
-    'XMLSchemaDecodeError', 'XMLSchemaEncodeError', 'XMLSchemaChildrenValidationError',
-    'XMLSchemaIncludeWarning', 'XMLSchemaImportWarning', 'XMLSchemaTypeTableWarning',
+    'JsonMLConverter', 'ColumnarConverter', 'validate', 'is_valid', 'iter_errors',
+    'to_dict', 'to_json', 'from_json', 'XMLSchemaValidatorError', 'XMLSchemaParseError',
+    'XMLSchemaNotBuiltError', 'XMLSchemaModelError', 'XMLSchemaModelDepthError',
+    'XMLSchemaValidationError', 'XMLSchemaDecodeError', 'XMLSchemaEncodeError',
+    'XMLSchemaChildrenValidationError', 'XMLSchemaIncludeWarning',
+    'XMLSchemaImportWarning', 'XMLSchemaTypeTableWarning',
     'XsdGlobals', 'XMLSchemaBase', 'XMLSchema', 'XMLSchema10', 'XMLSchema11',
+    'XsdComponent', 'XsdType', 'XsdElement', 'XsdAttribute',
 ]