+++ xmlschema/validators/simple_types.py
@@ -293,6 +293,12 @@ class XsdSimpleType(XsdType, ValidationMixin):
             return max_inclusive_facet.value
 
     @property
+    def enumeration(self):
+        enumeration = self.get_facet(XSD_ENUMERATION)
+        if enumeration is not None:
+            return enumeration.enumeration
+
+    @property
     def admitted_facets(self):
         return XSD_10_FACETS if self.xsd_version == '1.0' else XSD_11_FACETS
 
@@ -308,10 +314,6 @@ class XsdSimpleType(XsdType, ValidationMixin):
     def is_complex():
         return False
 
-    @staticmethod
-    def is_list():
-        return False
-
     def is_empty(self):
         return self.max_length == 0
 
@@ -320,6 +322,9 @@ class XsdSimpleType(XsdType, ValidationMixin):
 
     def has_simple_content(self):
         return True
+
+    def has_complex_content(self):
+        return False
 
     def has_mixed_content(self):
         return False
@@ -356,10 +361,9 @@ class XsdSimpleType(XsdType, ValidationMixin):
     def normalize(self, text):
         """
         Normalize and restrict value-space with pre-lexical and lexical facets.
-        The normalized string is returned. Returns the argument if it isn't a string.
 
         :param text: text string encoded value.
-        :return: normalized string.
+        :return: a normalized string.
         """
         if isinstance(text, bytes):
             text = text.decode('utf-8')
@@ -577,6 +581,14 @@ class XsdAtomicBuiltin(XsdAtomic):
                                 yield self.validation_error(validation, error=reason, obj=obj)
                         except KeyError:
                             pass
+            else:
+                try:
+                    default_namespace = kwargs['namespaces']['']
+                except (TypeError, KeyError):
+                    pass
+                else:
+                    if default_namespace:
+                        result = '{%s}%s' % (default_namespace, obj)
 
         elif self.name == XSD_IDREF:
             try:
@@ -922,7 +934,7 @@ class XsdUnion(XsdSimpleType):
 
     def is_dynamic_consistent(self, other):
         return other.name in (XSD_ANY_TYPE, XSD_ANY_SIMPLE_TYPE) or \
-            other.is_derived(self) or hasattr(other, 'member_types') and \
+            other.is_derived(self) or isinstance(other, self.__class__) and \
             any(mt1.is_derived(mt2) for mt1 in other.member_types for mt2 in self.member_types)
 
     def iter_components(self, xsd_classes=None):